Add getcape to isf
