+++ tests/.pytest_cache/v/cache/nodeids
@@ -23,5 +23,6 @@
   "test_idealized_sounding_fcts.py::test_get_es",
   "test_idealized_sounding_fcts.py::test_get_qvs",
   "test_getcape.py::test_getcape_pseudo",
-  "test_getcape.py::test_getcape"
+  "test_getcape.py::test_getcape",
+  "test_idealized_sounding_fcts.py::test_getcape"
 ]
